Refactor Stock numeric fields from long to BigDecimal to eliminate overflow
All price fields (currentPrice, basePrice, listingPrice) and accumulation
fields (dailyVolume, dailyTradingValue, issuedShares, totalSupply,
preStreamFloat) converted from long/BIGINT to BigDecimal/DECIMAL(65,x).
Removes all saturatingAdd/toLongCap workarounds from these fields.
V60 migration widens the corresponding DB columns.

+++ backend/src/main/java/com/spotchzxk/application/CandleService.java
@@ -119,7 +119,7 @@ public class CandleService {
             // If a trade occurred this minute, it was already broadcast by onTrade ??skip
             if (tradedStockIds.contains(stockId)) continue;
 
-            double price = stock.getCurrentPrice();
+            double price = stock.getCurrentPrice().doubleValue();
             Map<String, OhlcCandle> update = new HashMap<>();
             for (String interval : ALL_INTERVALS) {
                 long bms = INTERVAL_MS.get(interval);
